Add files via upload
This is folder of maven project. Main programs are Wyszukiwarka and Indekser.
Wyszukiwarka:
if command starts with %:
%lang pl/en - changes language
%limit <n> - sets max number of results
%details on/off - show a context and highlights a key word
%color on\off - if details on, it color a key word
%term - looks for the term(TermQuery)
%phrase - looks for the phrase(PhraseQuery)
%fuzzy - looks for the fuzzy text(FuzzyQuery)
else:
searches for the string given as argument in index. 

Indekser:
creates index, updates index in case of some changes

it can run with some arguments:
--add <dir>
marks this directory to the set of indexed directories
--rm <dir>
delete this dir from the set of indexed directories
--reindex
refresh the index, so that there are only these files, that exists a dir in the set of indexed directories, so that a file belongs to dir.
--list 
lists the set of indexed directories
(without arguments)
watches indexed directories and updates index after some changes in these directories.
Finishes, when ctrl+C is pressed.